% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{segment_reactor_output}
\alias{segment_reactor_output}
\title{Segment Shinra reactor data}
\usage{
segment_reactor_output(reactor_num, data_table = FALSE)
}
\arguments{
\item{reactor_num}{The reactor number to segment by.}

\item{data_table}{}
}
\value{
a tibble or data.table filtered by \code{reactor_num}
}
\description{
Segment Shinra reactor data
}
\examples{

segment_reactor_output(7)

}
